IN theory ready for disease model calibration.
Added the external scripts used for calculating the radius
of gyration and the mobility index.

Updated the code to account for the slight changes in calling the Java behavior model,
specifically, a state file is no longer used. Instead, a small number of agents is
infected for a small number of days. Both these values are now part of the disease
model calibration.

Added CSV-format logging of results, to make it easier to find the best run.