Made calibration code compatible with new 2APL + new run configuration
2APL now outputs the number of agents the radius of gyration was calculated
on. This allows agents of the same county to be distributed across multiple
compute nodes, resulting in the (average) radius of gyration being stored
across multiple tick_averages files.

The calibration code now is able to calculate the actual average over a
county, based on the averages of the individual compute nodes.

Added a run configuration to test the most extreme trust parameters,
to study the sensitivity of the simulation to trust. This configuration
just runs 5 simulations with the 4 pairs of most extreme trust parameters,
and a simulation with the trust of both agent groups set to 0.5