Added new data set concerning cases in Virginia.
The data was provided by the covid19-projections.com project.
The data has been rewritten to the same format as the previous
case data file.

Small updates to the code to support working with the new case
data have been made.

An analysis script for case data has been added.